requirements.txt
# Core Airflow dependencies
apache-airflow==2.6.1
apache-airflow-providers-snowflake==4.0.2
apache-airflow-providers-docker==3.5.0
apache-airflow-providers-apache-spark==4.0.0
apache-airflow-providers-http==4.1.0

# Common data processing libraries
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2

# Database connectors
snowflake-connector-python==3.0.2
pymysql==1.0.3
psycopg2-binary==2.9.5

# API integrations
requests==2.28.2
httpx==0.23.3

# Testing libraries
pytest==7.3.1
pytest-mock==3.10.0

# Development tools
black==23.3.0
flake8==6.0.0
pylint==2.17.0
